+++ src/UriHelper.php
@@ -19,16 +19,17 @@ class UriHelper
      *
      * @see UriHelper::addOptionsToUri
      *
-     * @param string       $url     The rokka image render URL
-     * @param array|string $options The options you want to add as string
+     * @param string       $url        The rokka image render URL
+     * @param array|string $options    The options you want to add as string
+     * @param bool         $shortNames if short names (like o for option or v for variables) should be used
      *
      * @throws \RuntimeException
      *
      * @return string
      */
-    public static function addOptionsToUriString($url, $options)
+    public static function addOptionsToUriString($url, $options, $shortNames = true)
     {
-        return (string) self::addOptionsToUri(new Uri($url), $options);
+        return (string) self::addOptionsToUri(new Uri($url), $options, $shortNames);
     }
 
     /**
@@ -61,17 +62,18 @@ class UriHelper
      * ```
      *
      *
-     * @param UriInterface $uri     The rokka image render URL
-     * @param array|string $options The options you want to add as string
+     * @param UriInterface $uri        The rokka image render URL
+     * @param array|string $options    The options you want to add as string
+     * @param bool         $shortNames if short names (like o for option or v for variables) should be used
      *
      * @throws \RuntimeException
      *
      * @return UriInterface
      */
-    public static function addOptionsToUri(UriInterface $uri, $options)
+    public static function addOptionsToUri(UriInterface $uri, $options, $shortNames = true)
     {
         if (\is_array($options)) {
-            return self::addOptionsToUri($uri, self::getUriStringFromStackConfig($options));
+            return self::addOptionsToUri($uri, self::getUriStringFromStackConfig($options, $shortNames), $shortNames);
         }
         $matches = self::decomposeUri($uri);
         if (empty($matches)) {
@@ -81,7 +83,7 @@ class UriHelper
         $stack = $matches->getStack();
         $stack->addOverridingOptions($options);
 
-        return self::composeUri($matches, $uri);
+        return self::composeUri($matches, $uri, $shortNames);
     }
 
     /**
@@ -100,12 +102,13 @@ class UriHelper
      *
      * @param array|UriComponents $components
      * @param UriInterface        $uri        If this is provided, it will change the path for that object and return
+     * @param bool                $shortNames if short names (like o for option or v for variables) should be used
      *
      * @throws \RuntimeException
      *
      * @return UriInterface
      */
-    public static function composeUri($components, UriInterface $uri = null)
+    public static function composeUri($components, UriInterface $uri = null, $shortNames = true)
     {
         if (\is_array($components)) {
             $components = UriComponents::createFromArray($components);
@@ -114,7 +117,7 @@ class UriHelper
         $stackName = $stack->getName();
         $path = '/'.$stackName;
         $stackConfig = $stack->getConfigAsArray();
-        $stackUrl = self::getUriStringFromStackConfig($stackConfig);
+        $stackUrl = self::getUriStringFromStackConfig($stackConfig, $shortNames);
         if (!empty($stackUrl)) {
             $path .= '/'.$stackUrl;
         }
@@ -167,17 +170,18 @@ class UriHelper
     }
 
     /**
-     * @param string      $url    The original rokka render URL to be adjusted
-     * @param string      $size   The size of the image, eg '300w' or '2x'
-     * @param null|string $custom Any rokka options you'd like to add, or are a dpi identifier like '2x'
+     * @param string      $url           The original rokka render URL to be adjusted
+     * @param string      $size          The size of the image, eg '300w' or '2x'
+     * @param null|string $custom        Any rokka options you'd like to add, or are a dpi identifier like '2x'
+     * @param bool        $setWidthInUrl If false, don't set the width as stack operation option, we provide it in $custom, usually as parameter
      *
      * @throws \RuntimeException
      *
      * @return UriInterface
      */
-    public static function getSrcSetUrlString($url, $size, $custom = null)
+    public static function getSrcSetUrlString($url, $size, $custom = null, $setWidthInUrl = true)
     {
-        return self::getSrcSetUrl(new Uri($url), $size, $custom);
+        return self::getSrcSetUrl(new Uri($url), $size, $custom, $setWidthInUrl);
     }
 
     /**
@@ -189,28 +193,34 @@ class UriHelper
      * This method will then generate the right rokka URLs to get what you want, see
      * `\Rokka\Client\Tests\UriHelperTest::provideGetSrcSetUrl` for some examples and the expected returns.
      *
-     * @param UriInterface $url    The original rokka render URL to be adjusted
-     * @param string       $size   The size of the image, eg '300w' or '2x'
-     * @param null|string  $custom Any rokka options you'd like to add, or are a dpi identifier like '2x'
+     * @param UriInterface $url           The original rokka render URL to be adjusted
+     * @param string       $size          The size of the image, eg '300w' or '2x'
+     * @param null|string  $custom        Any rokka options you'd like to add, or are a dpi identifier like '2x'
+     * @param bool         $setWidthInUrl If false, don't set the width as stack operation option, we provide it in $custom, usually as parameter
      *
      * @throws \RuntimeException
      *
      * @return UriInterface
      */
-    public static function getSrcSetUrl(UriInterface $url, $size, $custom = null)
+    public static function getSrcSetUrl(UriInterface $url, $size, $custom = null, $setWidthInUrl = true)
     {
         $identifier = substr($size, -1, 1);
         $size = substr($size, 0, -1);
         if ('x' === $identifier) {
             $uri = self::addOptionsToUri($url, 'options-dpr-'.$size);
         } elseif ('w' === $identifier) {
-            $uri = self::addOptionsToUri($url, 'resize-width-'.$size);
+            if ($setWidthInUrl) {
+                $uri = self::addOptionsToUri($url, 'resize-width-'.$size);
+            } else {
+                $uri = $url;
+            }
         } else {
             return $url;
         }
         if (null !== $custom) {
             if (preg_match('#^([0-9]+)x$#', $custom, $matches)) {
-                $uri = self::addOptionsToUri($uri, 'options-dpr-'.$matches[1].'--resize-width-'.(int) ceil($size / $matches[1]));
+                $uri = self::addOptionsToUri($uri, 'options-dpr-'.$matches[1].
+                    ($setWidthInUrl ? '--resize-width-'.(int) ceil($size / $matches[1]) : ''));
             } else {
                 $stack = new StackUri();
                 $stack->addOverridingOptions($custom);
@@ -223,7 +233,7 @@ class UriHelper
                     }
                 }
                 $options = $stack->getStackOptions();
-                if (isset($options['dpr']) && $widthIsNotSet) {
+                if (isset($options['dpr']) && $widthIsNotSet && $setWidthInUrl) {
                     $custom .= '--resize-width-'.(int) ceil($size / $options['dpr']);
                 }
 
@@ -236,10 +246,11 @@ class UriHelper
 
     /**
      * @param array $config
+     * @param bool  $shortNames if short names (like o for option or v for variables) should be used
      *
      * @return string
      */
-    private static function getUriStringFromStackConfig(array $config)
+    private static function getUriStringFromStackConfig(array $config, $shortNames = true)
     {
         $newOptions = [];
 
@@ -255,21 +266,24 @@ class UriHelper
                 }
             }
         }
+
         $newStackOptions = null;
+        $nameOptions = $shortNames ? 'o' : 'options';
         if (isset($config['options'])) {
-            $newStackOptions = self::getStringForOptions('options', $config['options']);
+            $newStackOptions = self::getStringForOptions($nameOptions, $config['options']);
         }
         //don't return this, if it's only "options" as string
-        if ('options' !== $newStackOptions) {
+        if (null !== $newStackOptions && $nameOptions !== $newStackOptions) {
             $newOptions[] = $newStackOptions;
         }
 
         $newStackVariables = null;
+        $nameVariables = $shortNames ? 'v' : 'variables';
         if (isset($config['variables'])) {
-            $newStackVariables = self::getStringForOptions('variables', $config['variables']);
+            $newStackVariables = self::getStringForOptions($nameVariables, $config['variables']);
         }
         //don't return this, if it's only "variables" as string
-        if ('variables' !== $newStackVariables && null !== $newStackVariables) {
+        if (null !== $newStackVariables && $nameVariables !== $newStackVariables) {
             $newOptions[] = $newStackVariables;
         }
         $options = implode('--', $newOptions);