+++ CHANGELOG.md
@@ -9,6 +9,12 @@ This document mainly describes API changes important to users of this library.
 ## 1.10.0 - unreleased
 
 * removed HHVM support on travis tests, therefore HHVM isn't officially supported anymore.
+* Added stack variables and expressions support. See [the documentation](/documentation/references/stacks.html#expressions) for more details.
+* Added optional 3rd boolean parameter to `\Rokka\Client\UriHelper::addOptionsToUriString` and 
+  related methods to return short versions for `options` (`o`) and `variables` (`v`)
+* Added optional 3rd boolean parameter to `\Rokka\Client\TemplateHelper::getSrcAttributes` to set
+  the stack operation option `width` for `resize`. If false, you should add that via the `$sizes`
+  parameter, eg `getSrcAttributes(['2x' => 'v-w-500]` (if you want to set the width via a stack variable)
 
 ## 1.9.0 - 2018-12-13